+++ public/index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <title>Mj</title>
+  <title>Home</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <!-- <meta name="description" content="Premium Bootstrap 5 Landing Page Template" />
   <meta name="keywords" content="Saas, Software, multi-uses, HTML, Clean, Modern" />
