+++ src/components/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react";
-import Navbar from "../../common/navbar/Navbar";
-import Footer from "../../common/footer/Footer";
+import Navbar from "../common/navbar/Navbar";
+import Footer from "../common/footer/Footer";
 
 interface LayoutProps {
   children: ReactNode;
