+++ src/components/auth/signup/ImageSection.tsx
@@ -1,13 +1,13 @@
-import React from 'react';
+import React from 'react'
 
-import loginImg from "../../assets/images/user/login.svg";
+import signup from "../../../assets/images/user/signup.svg";
 
 const ImageSection = () => {
   return (
     <React.Fragment>
       <div className="me-lg-5">
                 <img
-                  src={loginImg}
+                  src={signup}
                   className="img-fluid d-block mx-auto"
                   alt=""
                 />
