+++ src/components/common/footer/Footer.tsx
@@ -6,7 +6,7 @@ import { Col, Container, Row } from "reactstrap";
 import { Icon } from "ts-react-feather-icons";
 
 //Import Images
-import logoLight from "../../assets/images/logo-light.png";
+
 
 const Footer = () => {
   return (
@@ -17,15 +17,15 @@ const Footer = () => {
             <Row className="align-items-center">
               <Col sm={3}>
                 <div className="text-sm-start">
-                  <Link to="#" className="logo-footer">
-                    <img src={logoLight} height="24" alt="" />
+                  <Link to="/" className="logo-footer text-light">
+                    Mj Attar Shop
                   </Link>
                 </div>
               </Col>
               <Col sm={6} className="mt-4 mt-sm-0 pt-2 pt-sm-0">
                 <div className="text-center">
                   <p className="mb-0">
-                    © {new Date().getFullYear()} Landrick. Design with{" "}
+                    © {new Date().getFullYear()} Mj Attar Shop. Design with{" "}
                     <i className="mdi mdi-heart text-danger"></i> by{" "}
                     <Link
                       to="//themesbrand.com//"
@@ -33,7 +33,7 @@ const Footer = () => {
                       rel="noreferrer"
                       className="text-reset"
                     >
-                      Themesbrand
+                      MJ
                     </Link>
                     .
                   </p>
@@ -43,7 +43,7 @@ const Footer = () => {
               <Col sm={3} className="mt-4 mt-sm-0 pt-2 pt-sm-0">
                 <ul className="list-unstyled social-icon foot-social-icon text-sm-end mb-0">
                   <li className="list-inline-item mb-0">
-                    <Link to="#" className="rounded">
+                    <Link to={{pathname:"https://www.facebook.com/apexmustafamj"}} className="rounded" target="_blank">
                       <Icon
                         name="facebook"
                         
@@ -51,7 +51,7 @@ const Footer = () => {
                     </Link>
                   </li>{" "}
                   <li className="list-inline-item mb-0">
-                    <Link to="#" className="rounded">
+                    <Link to={{pathname:"https://www.instagram.com/mohammad_mustafa_mj/"}} className="rounded" target="_blank">
                       <Icon
                         name="instagram"
                        
@@ -59,7 +59,7 @@ const Footer = () => {
                     </Link>
                   </li>{" "}
                   <li className="list-inline-item mb-0">
-                    <Link to="#" className="rounded">
+                    <Link to={{pathname:"https://twitter.com/Apex_Mj"}} className="rounded" target="_blank">
                       <Icon
                         name="twitter"
                        
@@ -67,7 +67,7 @@ const Footer = () => {
                     </Link>
                   </li>{" "}
                   <li className="list-inline-item mb-0">
-                    <Link to="#" className="rounded">
+                    <Link to={{pathname:"https://www.linkedin.com/in/mustafa-jamadar-51b722232/"}} className="rounded" target="_blank">
                       <Icon
                         name="linkedin"
